move out replication and syncer
