+++ pkg/syncer/client_test.go
